+++ sunkit_instruments/iris/tests/test_iris_utils.py
@@ -11,9 +11,9 @@
 # import scipy.io
 # from sunpy.time import parse_time
 #
-# from sunkit_instruments.iris import iris_tools
-# from iris_tools import fit_iris_xput
-# from iris_tools import get_iris_response
+# from sunkit_instruments.iris.utils import iris_utils
+# from iris_utils import fit_iris_xput
+# from iris_utils import get_iris_response
 #
 #
 # DETECTOR_TYPE_KEY = "detector type"
@@ -67,28 +67,28 @@
 #     ({DETECTOR_TYPE_KEY: "SJI"}, "SJI")
 # ])
 # def test_get_detector_type(test_input, expected_output):
-#     assert iris_tools.get_detector_type(test_input) == expected_output
+#     assert iris_utils.get_detector_type(test_input) == expected_output
 #
 #
 # @pytest.mark.parametrize("data_arrays, old_unit, new_unit, expected_data_arrays, expected_unit",
 #                          [(
-#                                  [SOURCE_DATA_DN, SOURCE_DATA_DN], iris_tools.DN_UNIT["FUV"], u.photon,
+#                                  [SOURCE_DATA_DN, SOURCE_DATA_DN], iris_utils.DN_UNIT["FUV"], u.photon,
 #                                  [SOURCE_DATA_PHOTONS_FUV, SOURCE_DATA_PHOTONS_FUV], u.photon),
-#                              ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_tools.DN_UNIT["NUV"], u.photon,
+#                              ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_utils.DN_UNIT["NUV"], u.photon,
 #                               [SOURCE_DATA_PHOTONS_NUV, SOURCE_DATA_PHOTONS_NUV], u.photon),
-#                              ([SOURCE_DATA_DN_1, SOURCE_DATA_DN_1], iris_tools.DN_UNIT["SJI"], u.photon,
+#                              ([SOURCE_DATA_DN_1, SOURCE_DATA_DN_1], iris_utils.DN_UNIT["SJI"], u.photon,
 #                               [SOURCE_DATA_PHOTONS_SJI_1, SOURCE_DATA_PHOTONS_SJI_1], u.photon),
-#                              ([SOURCE_DATA_PHOTONS_FUV, SOURCE_DATA_PHOTONS_FUV], u.photon, iris_tools.DN_UNIT["FUV"],
-#                               [SOURCE_DATA_DN, SOURCE_DATA_DN], iris_tools.DN_UNIT["FUV"]),
-#                              ([SOURCE_DATA_PHOTONS_NUV, SOURCE_DATA_PHOTONS_NUV], u.photon, iris_tools.DN_UNIT["NUV"],
-#                               [SOURCE_DATA_DN, SOURCE_DATA_DN], iris_tools.DN_UNIT["NUV"]),
-#                              ([SOURCE_DATA_PHOTONS_SJI_1, SOURCE_DATA_PHOTONS_SJI_1], u.photon, iris_tools.DN_UNIT["SJI"],
-#                               [SOURCE_DATA_DN_1, SOURCE_DATA_DN_1], iris_tools.DN_UNIT["SJI"]
+#                              ([SOURCE_DATA_PHOTONS_FUV, SOURCE_DATA_PHOTONS_FUV], u.photon, iris_utils.DN_UNIT["FUV"],
+#                               [SOURCE_DATA_DN, SOURCE_DATA_DN], iris_utils.DN_UNIT["FUV"]),
+#                              ([SOURCE_DATA_PHOTONS_NUV, SOURCE_DATA_PHOTONS_NUV], u.photon, iris_utils.DN_UNIT["NUV"],
+#                               [SOURCE_DATA_DN, SOURCE_DATA_DN], iris_utils.DN_UNIT["NUV"]),
+#                              ([SOURCE_DATA_PHOTONS_SJI_1, SOURCE_DATA_PHOTONS_SJI_1], u.photon, iris_utils.DN_UNIT["SJI"],
+#                               [SOURCE_DATA_DN_1, SOURCE_DATA_DN_1], iris_utils.DN_UNIT["SJI"]
 #                               )]
 #                          )
 # def test_convert_between_DN_and_photons(data_arrays, old_unit, new_unit,
 #                                         expected_data_arrays, expected_unit):
-#     output_arrays, output_unit = iris_tools.convert_between_DN_and_photons(data_arrays,
+#     output_arrays, output_unit = iris_utils.convert_between_DN_and_photons(data_arrays,
 #                                                                            old_unit, new_unit)
 #     for i, output_array in enumerate(output_arrays):
 #         np_test.assert_allclose(output_array, expected_data_arrays[i])
@@ -100,19 +100,19 @@
 #         ([SOURCE_DATA_DN, SOURCE_DATA_DN], u.photon, EXPOSURE_TIME, False,
 #          [SOURCE_DATA_DN/single_exposure_time, SOURCE_DATA_DN/single_exposure_time],
 #          u.photon/u.s),
-#         ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_tools.DN_UNIT["NUV"], EXPOSURE_TIME, False,
+#         ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_utils.DN_UNIT["NUV"], EXPOSURE_TIME, False,
 #          [SOURCE_DATA_DN/single_exposure_time, SOURCE_DATA_DN/single_exposure_time],
-#          iris_tools.DN_UNIT["NUV"]/u.s),
+#          iris_utils.DN_UNIT["NUV"]/u.s),
 #         ([SOURCE_DATA_DN, SOURCE_DATA_DN], u.photon/u.s, EXPOSURE_TIME, True,
 #          [SOURCE_DATA_DN/single_exposure_time, SOURCE_DATA_DN/single_exposure_time],
 #          u.photon/u.s/u.s),
-#         ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_tools.DN_UNIT["NUV"]/u.s, EXPOSURE_TIME, True,
+#         ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_utils.DN_UNIT["NUV"]/u.s, EXPOSURE_TIME, True,
 #          [SOURCE_DATA_DN/single_exposure_time, SOURCE_DATA_DN/single_exposure_time],
-#          iris_tools.DN_UNIT["NUV"]/u.s/u.s)
+#          iris_utils.DN_UNIT["NUV"]/u.s/u.s)
 #     ])
 # def test_calculate_exposure_time_correction(input_arrays, old_unit, exposure_time, force,
 #                                             expected_arrays, expected_unit):
-#     output_arrays, output_unit = iris_tools.calculate_exposure_time_correction(
+#     output_arrays, output_unit = iris_utils.calculate_exposure_time_correction(
 #         input_arrays, old_unit, exposure_time, force=force)
 #     for i, output_array in enumerate(output_arrays):
 #         np_test.assert_allclose(output_array, expected_arrays[i])
@@ -121,10 +121,10 @@
 #
 # @pytest.mark.parametrize("input_arrays, old_unit, exposure_time, force", [
 #     ([SOURCE_DATA_DN, SOURCE_DATA_DN], u.photon/u.s, EXPOSURE_TIME, False),
-#     ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_tools.DN_UNIT["NUV"]/u.s, EXPOSURE_TIME, False)
+#     ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_utils.DN_UNIT["NUV"]/u.s, EXPOSURE_TIME, False)
 # ])
 # def test_calculate_exposure_time_correction_error(input_arrays, old_unit, exposure_time, force):
-#     assert pytest.raises(ValueError, iris_tools.calculate_exposure_time_correction,
+#     assert pytest.raises(ValueError, iris_utils.calculate_exposure_time_correction,
 #                          input_arrays, old_unit, exposure_time, force)
 #
 #
@@ -132,19 +132,19 @@
 #     "input_arrays, old_unit, exposure_time, force, expected_arrays, expected_unit", [
 #         ([SOURCE_DATA_DN, SOURCE_DATA_DN], u.photon/u.s, EXPOSURE_TIME, False,
 #          [SOURCE_DATA_DN * single_exposure_time, SOURCE_DATA_DN * single_exposure_time], u.photon),
-#         ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_tools.DN_UNIT["NUV"]/u.s, EXPOSURE_TIME, False,
+#         ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_utils.DN_UNIT["NUV"]/u.s, EXPOSURE_TIME, False,
 #          [SOURCE_DATA_DN * single_exposure_time, SOURCE_DATA_DN * single_exposure_time],
-#          iris_tools.DN_UNIT["NUV"]),
+#          iris_utils.DN_UNIT["NUV"]),
 #         ([SOURCE_DATA_DN, SOURCE_DATA_DN], u.photon, EXPOSURE_TIME, True,
 #          [SOURCE_DATA_DN * single_exposure_time, SOURCE_DATA_DN * single_exposure_time],
 #          u.photon*u.s),
-#         ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_tools.DN_UNIT["FUV"], EXPOSURE_TIME, True,
+#         ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_utils.DN_UNIT["FUV"], EXPOSURE_TIME, True,
 #          [SOURCE_DATA_DN * single_exposure_time, SOURCE_DATA_DN * single_exposure_time],
-#          iris_tools.DN_UNIT["FUV"]*u.s)
+#          iris_utils.DN_UNIT["FUV"]*u.s)
 #     ])
 # def test_uncalculate_exposure_time_correction(input_arrays, old_unit, exposure_time, force,
 #                                               expected_arrays, expected_unit):
-#     output_arrays, output_unit = iris_tools.uncalculate_exposure_time_correction(
+#     output_arrays, output_unit = iris_utils.uncalculate_exposure_time_correction(
 #         input_arrays, old_unit, exposure_time, force=force)
 #     for i, output_array in enumerate(output_arrays):
 #         np_test.assert_allclose(output_array, expected_arrays[i])
@@ -153,21 +153,21 @@
 #
 # @pytest.mark.parametrize("input_arrays, old_unit, exposure_time, force", [
 #     ([SOURCE_DATA_DN, SOURCE_DATA_DN], u.photon, EXPOSURE_TIME, False),
-#     ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_tools.DN_UNIT["NUV"], EXPOSURE_TIME, False)
+#     ([SOURCE_DATA_DN, SOURCE_DATA_DN], iris_utils.DN_UNIT["NUV"], EXPOSURE_TIME, False)
 # ])
 # def test_uncalculate_exposure_time_correction_error(input_arrays, old_unit, exposure_time, force):
 #     with pytest.raises(ValueError):
-#         assert iris_tools.uncalculate_exposure_time_correction(input_arrays, old_unit,
+#         assert iris_utils.uncalculate_exposure_time_correction(input_arrays, old_unit,
 #                                                                exposure_time, force=force)
 #
 #
 # def test_get_iris_response_not_equal_to_one():
-#     assert pytest.raises(KeyError, iris_tools.get_iris_response, time_obs,
+#     assert pytest.raises(KeyError, iris_utils.get_iris_response, time_obs,
 #                          pre_launch=False, response_version=13)
 #
 #
 # def test_get_iris_response_response_file():
-#     assert pytest.raises(KeyError, iris_tools.get_iris_response, time_obs, response_file="hello.py")
+#     assert pytest.raises(KeyError, iris_utils.get_iris_response, time_obs, response_file="hello.py")
 #
 #
 # # Tests for get_iris_response function
@@ -306,7 +306,7 @@
 # @pytest.mark.parametrize("input_array, expected_array", [
 #     (data_dust, dust_mask_expected)])
 # def test_calculate_dust_mask(input_array, expected_array):
-#     np_test.assert_array_equal(iris_tools.calculate_dust_mask(input_array), expected_array)
+#     np_test.assert_array_equal(iris_utils.calculate_dust_mask(input_array), expected_array)
 #
 #
 # @pytest.mark.parametrize("input_arrays, expected_array", [
